// iron/src/app/interfaces/gerentes_proyectos.ts
import { Title } from '@angular/platform-browser';

export interface Gerentes_proyectos{
   	
         gp_id?: number;
         gp_nombres:string;
		 gp_apellidos: string;
		 gp_correo: string;
		 gp_celular_coorporativo: string;
		 gp_celular_personal: string;
		 gp_cliente : string;
		 gp_centro_costos : string;
}